requirements.txt
fastapi
uvicorn
pydantic
google-genai
python-jose[cryptography]==3.5.0